pyproject.toml
[tool.poetry]
name = "shoponsmile-python"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
requests-oauthlib = "^1.3.0"
python-dotenv = "^0.13.0"
tweepy = "^3.8.0"
oauth2 = "^1.9.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
